+++ BearoundSDK/Models/SDKInfo.swift
@@ -13,7 +13,7 @@ struct SDKInfo {
     let appId: String
     let build: Int
 
-    init(version: String = "2.2.0", appId: String, build: Int) {
+    init(version: String = "2.2.1", appId: String, build: Int) {
         self.version = version
         platform = "ios"
         self.appId = appId